+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 * Fix an issue with image caption updating that may replace text surrounding the
   image (#5021)
 * Too small images won't display the informational layers anymore (#3953)
+* Blockquote markers are now properly hidden with the emphasis renderer on and
+  the cursor not within the blockquote (#4667)
 
 ## Under the Hood